bugfix (取りきれた？)

+++ src/main/java/xyz/n7mn/dev/yululi/itemframeprotectionplugin/ItemFrameTimer.java
@@ -1,6 +1,7 @@
 package xyz.n7mn.dev.yululi.itemframeprotectionplugin;
 
 import org.bukkit.ChatColor;
+import org.bukkit.World;
 import org.bukkit.event.Cancellable;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -32,7 +33,7 @@ class ItemFrameTimer extends BukkitRunnable implements Cancellable {
             if (con != null){
                 try {
                     List<FrameData> itemFrameList = data.getItemFrameList();
-                    System.out.println("Debug : LockFrame : " + itemFrameList.size());
+                    // System.out.println("Debug : LockFrame : " + itemFrameList.size());
                     for (FrameData data : itemFrameList){
                         PreparedStatement statement = con.prepareStatement("SELECT COUNT(*) FROM IFPTable WHERE ItemFrame = ?");
                         statement.setString(1, data.getItemFrame().toString());
@@ -50,7 +51,7 @@ class ItemFrameTimer extends BukkitRunnable implements Cancellable {
                     }
 
                     List<DropData> dropList = data.getDropList();
-                    System.out.println("Debug : dropItem : " + dropList.size());
+                    // System.out.println("Debug : dropItem : " + dropList.size());
                     for (DropData data : dropList){
                         PreparedStatement statement = con.prepareStatement("SELECT COUNT(*) FROM IFPTable2 WHERE ItemUUID = ?");
                         statement.setString(1, data.getItemUUID().toString());
@@ -68,33 +69,41 @@ class ItemFrameTimer extends BukkitRunnable implements Cancellable {
                     }
 
                     // ここからゴミデータお掃除
-                    PreparedStatement statement = con.prepareStatement("SELECT * FROM IFPTable");
-                    ResultSet set = statement.executeQuery();
-                    while(set.next()){
-                        FrameData itemFrame = data.getItemFrame(UUID.fromString(set.getString("CreateUser")), UUID.fromString(set.getString("ItemFrame")));
-                        if (itemFrame == null){
-                            PreparedStatement statement1 = con.prepareStatement("DELETE FROM `IFPTable` WHERE `CreateUser` = ? AND `ItemFrame` = ?");
-                            statement1.setString(1, set.getString("CreateUser"));
-                            statement1.setString(2, set.getString("ItemFrame"));
-                            statement1.execute();
-                            statement1.close();
-                        }
-                    }
-                    statement.close();
 
-                    PreparedStatement statement2 = con.prepareStatement("SELECT * FROM IFPTable2");
-                    ResultSet set2 = statement2.executeQuery();
-                    while(set2.next()){
-                        FrameData itemFrame = data.getItemFrame(UUID.fromString(set.getString("DropUser")), UUID.fromString(set.getString("ItemUUID")));
-                        if (itemFrame == null){
-                            PreparedStatement statement1 = con.prepareStatement("DELETE FROM `IFPTable2` WHERE `DropUser` = ? AND `ItemUUID` = ?");
-                            statement1.setString(1, set.getString("DropUser"));
-                            statement1.setString(2, set.getString("ItemUUID"));
-                            statement1.execute();
-                            statement1.close();
-                        }
+                    List<World> worlds = plugin.getServer().getWorlds();
+                    int i = 0;
+                    for (World world : worlds){
+                        i = i + world.getEntities().size();
                     }
-                    statement2.close();
+                    if (i != 0){
+                        PreparedStatement statement = con.prepareStatement("SELECT * FROM IFPTable");
+                        ResultSet set = statement.executeQuery();
+                        while(set.next()){
+                            FrameData itemFrame = data.getItemFrame(UUID.fromString(set.getString("CreateUser")), UUID.fromString(set.getString("ItemFrame")));
+                            if (itemFrame == null){
+                                PreparedStatement statement1 = con.prepareStatement("DELETE FROM `IFPTable` WHERE `CreateUser` = ? AND `ItemFrame` = ?");
+                                statement1.setString(1, set.getString("CreateUser"));
+                                statement1.setString(2, set.getString("ItemFrame"));
+                                statement1.execute();
+                                statement1.close();
+                            }
+                        }
+                        statement.close();
+
+                        PreparedStatement statement2 = con.prepareStatement("SELECT * FROM IFPTable2");
+                        ResultSet set2 = statement2.executeQuery();
+                        while(set2.next()){
+                            FrameData itemFrame = data.getItemFrame(UUID.fromString(set.getString("DropUser")), UUID.fromString(set.getString("ItemUUID")));
+                            if (itemFrame == null){
+                                PreparedStatement statement1 = con.prepareStatement("DELETE FROM `IFPTable2` WHERE `DropUser` = ? AND `ItemUUID` = ?");
+                                statement1.setString(1, set.getString("DropUser"));
+                                statement1.setString(2, set.getString("ItemUUID"));
+                                statement1.execute();
+                                statement1.close();
+                            }
+                        }
+                        statement2.close();
+                    }
 
                 } catch (SQLException e){
                     if (plugin.getConfig().getBoolean("errorPrint")) {
